requirements.txt
aniso8601==8.0.0
click==7.1.1
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-MySQL==1.5.1
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dotenv==0.13.0
pytz==2019.3
six==1.14.0
Werkzeug==1.0.1
